Move files from src to double-conversion.Also update include to use the same include path like the users of double-conversion, i.e. include "double-conversion/util.h" instead of include "util.h". This is to make sure double-conversion src can be used without further operation(like symlink double-conversion to src) or install to /usr/include.
Tested by running make tests and bazel build.

Add "-lm" to linkopts of BUILD.
