+++ BUILD
@@ -6,28 +6,31 @@ cc_library(
   name = "double-conversion",
   visibility = ["//visibility:public"],
   hdrs = [
-    "src/bignum-dtoa.h",
-    "src/bignum.h",
-    "src/cached-powers.h",
-    "src/diy-fp.h",
-    "src/double-conversion.h",
-    "src/fast-dtoa.h",
-    "src/fixed-dtoa.h",
-    "src/ieee.h", 
-    "src/strtod.h",
+    "double-conversion/bignum-dtoa.h",
+    "double-conversion/bignum.h",
+    "double-conversion/cached-powers.h",
+    "double-conversion/diy-fp.h",
+    "double-conversion/double-conversion.h",
+    "double-conversion/fast-dtoa.h",
+    "double-conversion/fixed-dtoa.h",
+    "double-conversion/ieee.h", 
+    "double-conversion/strtod.h",
   ],
   srcs = [
-    "src/bignum-dtoa.cc",
-    "src/bignum.cc",
-    "src/cached-powers.cc",
-    "src/diy-fp.cc",
-    "src/double-conversion.cc",
-    "src/fast-dtoa.cc",
-    "src/fixed-dtoa.cc",
-    "src/strtod.cc",
-    "src/utils.h"
+    "double-conversion/bignum-dtoa.cc",
+    "double-conversion/bignum.cc",
+    "double-conversion/cached-powers.cc",
+    "double-conversion/diy-fp.cc",
+    "double-conversion/double-conversion.cc",
+    "double-conversion/fast-dtoa.cc",
+    "double-conversion/fixed-dtoa.cc",
+    "double-conversion/strtod.cc",
+    "double-conversion/utils.h"
   ],
   includes = [
-    "src",
+    ".",
   ],
+  linkopts = [
+    "-lm",
+  ]
 )
