+++ double-conversion/cached-powers.h
