+++ double-conversion/ieee.h
