+++ double-conversion/utils.h
