+++ test/cctest/cctest.h
@@ -31,7 +31,7 @@
 #include <stdio.h>
 #include <string.h>
 
-#include "utils.h"
+#include "double-conversion/utils.h"
 
 #ifndef TEST
 #define TEST(Name)                                                       \
